Fixed warnings
* Don't use signed and unsigned tyeps interchangeably
* Use intptr_t for smuggling integers in pointer types. This breaks
  compatibility with antique compilers and weird hardware where
  intptr_t is impossible, but we don't care about any of those.

+++ libhfs/os/unix.c
@@ -83,7 +83,7 @@ int os_open(void **priv, const char *path, int mode)
       (errno == EACCES || errno == EAGAIN))
     ERROR(EAGAIN, "unable to obtain lock for medium");
 
-  *priv = (void *) fd;
+  *priv = (void *) (intptr_t) fd;
 
   return 0;
 
@@ -100,7 +100,7 @@ fail:
  */
 int os_close(void **priv)
 {
-  int fd = (int) *priv;
+  int fd = (intptr_t) *priv;
 
   *priv = (void *) -1;
 
@@ -119,7 +119,7 @@ fail:
  */
 int os_same(void **priv, const char *path)
 {
-  int fd = (int) *priv;
+  int fd = (intptr_t) *priv;
   struct stat fdev, dev;
 
   if (fstat(fd, &fdev) == -1 ||
@@ -139,7 +139,7 @@ fail:
  */
 unsigned long os_seek(void **priv, unsigned long offset)
 {
-  int fd = (int) *priv;
+  int fd = (intptr_t) *priv;
   off_t result;
 
   /* offset == -1 special; seek to last block of device */
@@ -164,7 +164,7 @@ fail:
  */
 unsigned long os_read(void **priv, void *buf, unsigned long len)
 {
-  int fd = (int) *priv;
+  int fd = (intptr_t) *priv;
   ssize_t result;
 
   result = read(fd, buf, len << HFS_BLOCKSZ_BITS);
@@ -184,7 +184,7 @@ fail:
  */
 unsigned long os_write(void **priv, const void *buf, unsigned long len)
 {
-  int fd = (int) *priv;
+  int fd = (intptr_t) *priv;
   ssize_t result;
 
   result = write(fd, buf, len << HFS_BLOCKSZ_BITS);
